+++ packages/chains/chains/2222.ts
@@ -6,10 +6,22 @@ export default {
     {
       "name": "Kava EVM Explorer",
       "url": "https://kavascan.com",
-      "standard": "EIP3091"
+      "standard": "EIP3091",
+      "icon": {
+        "url": "ipfs://QmdpRTk6oL1HRW9xC6cAc4Rnf9gs6zgdAcr4Z3HcLztusm",
+        "width": 1186,
+        "height": 360,
+        "format": "svg"
+      }
     }
   ],
   "faucets": [],
+  "icon": {
+    "url": "ipfs://QmdpRTk6oL1HRW9xC6cAc4Rnf9gs6zgdAcr4Z3HcLztusm",
+    "width": 1186,
+    "height": 360,
+    "format": "svg"
+  },
   "infoURL": "https://www.kava.io",
   "name": "Kava",
   "nativeCurrency": {
@@ -25,7 +37,10 @@ export default {
     "https://kava-rpc.gateway.pokt.network",
     "wss://wevm.kava.io",
     "https://kava-evm.publicnode.com",
-    "wss://kava-evm.publicnode.com"
+    "wss://kava-evm.publicnode.com",
+    "https://evm.kava-rpc.com",
+    "https://rpc.ankr.com/kava_evm",
+    "wss://wevm.kava-rpc.com"
   ],
   "shortName": "kava",
   "slug": "kava",