+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @thirdweb-dev/react-native
 
+## 0.6.4
+
+### Patch Changes
+
+- [#2199](https://github.com/thirdweb-dev/js/pull/2199) [`57320957`](https://github.com/thirdweb-dev/js/commit/57320957aae56d063eca526a05df8bea8b38ba2a) Thanks [@iketw](https://github.com/iketw)! - Adds ability to delete the active ews account
+
+- [#2217](https://github.com/thirdweb-dev/js/pull/2217) [`7f35caff`](https://github.com/thirdweb-dev/js/commit/7f35caff5d9d331e66125bab3797d7ca84f6cef1) Thanks [@MananTank](https://github.com/MananTank)! - Update defaultTokens used in ConnectWallet component's Send Funds screen
+
+- Updated dependencies [[`a5653f07`](https://github.com/thirdweb-dev/js/commit/a5653f07f82e3f94e193d2b90c9d0c163223583c), [`b34737d2`](https://github.com/thirdweb-dev/js/commit/b34737d27a71706cd1117485a6e522e50f5ae4d5), [`bf90fa50`](https://github.com/thirdweb-dev/js/commit/bf90fa5005358e040475627ae23a474f7b263803), [`627d2099`](https://github.com/thirdweb-dev/js/commit/627d20992770670ab0c954ceea0b8679ce4f3088), [`3d63a2cb`](https://github.com/thirdweb-dev/js/commit/3d63a2cb696310507798889e3e148339de4e3dfe), [`6fdade6c`](https://github.com/thirdweb-dev/js/commit/6fdade6c45f5d93fb3571d33720ab8103a7d002c), [`4d89dc1e`](https://github.com/thirdweb-dev/js/commit/4d89dc1efa2412b1e121cd14b1e2e72c667fceb0)]:
+  - @thirdweb-dev/wallets@2.4.4
+  - @thirdweb-dev/sdk@4.0.29
+  - @thirdweb-dev/chains@0.1.66
+  - @thirdweb-dev/react-core@4.4.1
+
 ## 0.6.3
 
 ### Patch Changes